requirements.txt
# ============================================
# DeepTutor Python Dependencies
# ============================================
# Python >= 3.10 required
# Install: pip install -r requirements.txt
#
# NOTE: raganything (multi-modal RAG) is commented out due to complex
# dependencies (gradio, mlx-vlm, etc.). Install separately if needed:
#   pip install raganything
# ============================================

# ============================================
# Core dependencies
# ============================================
python-dotenv>=1.0.0
PyYAML>=6.0
tiktoken>=0.5.0

# ============================================
# HTTP and API clients
# ============================================
requests>=2.32.2
openai>=1.30.0
perplexityai>=0.1.0
exa-py>=1.2.0  # Exa AI for enhanced web search
dashscope>=1.14.0
aiohttp>=3.9.4
httpx>=0.27.0
urllib3>=2.2.1

# ============================================
# Async support
# ============================================
nest_asyncio>=1.5.8

# ============================================
# Web framework and server
# ============================================
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.0.0,<3.0.0
pydantic-core>=2.18.0,<3.0.0

# ============================================
# RAG and knowledge base
# ============================================
lightrag-hku>=1.0.0
# raganything>=0.1.0  # Optional: Multi-modal RAG (complex dependencies)
# Install separately if needed: pip install raganything

# ============================================
# Academic and research tools
# ============================================
arxiv>=2.0.0

# ============================================
# Scientific computing (optional, for code execution)
# ============================================
# Uncomment if users need these for code execution feature
# numpy>=1.24.0
# matplotlib>=3.7.0
# scipy>=1.11.0
# pandas>=2.0.0

# ============================================
# Development tools
# ============================================
pre-commit>=3.0.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.12
llama-index-cli==0.5.3
llama-index-core==0.14.12
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.12
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.6
llama-parse==0.6.54
